function_app/requirements.txt
azure-functions
pandas
requests
openpyxl
azure-storage-blob
python-dateutil